donnie_contrib/requirements.txt
discord.py==2.2.3
python-dotenv==1.0.0
aiohttp==3.8.5
pytz==2023.3
